requirements.txt
pypdf
langchain
sentence-transformers
faiss-cpu
sqlite-utils
whoosh
fastapi
uvicorn
python-multipart
scikit-learn
nltk